EnoFJS is now more elegantly integrated with node.js.
This is however a major change. The old version would just append the module to the global `window`. Now it is refactored to actually append the module on `window.module`.

Also the `ClassFactory` got renamed to `clazz` to represent the module exposure name.

The nodejs version will use the commonjs `require` to include modules.